VTKU-94 : Päivitä AWS-SDK
Tämä voisi selittää erikoisen S3-liikenteen ongelman,
joka tuli esiin testissä.

Mockito täytyy päivittää, koska jotkut testeissä mockatut
luokat ovat muuttuneet finaleiksi, ja final-luokkia voi
mockata vain Mockiton 2-versiolla tai uudemmalla.

Uusi Mockito (tai Powermock?) ei enää machaa any(Luokka.class)
-matcherila nulleihin argumentteihin.

+++ viestintapalvelu-service/src/test/java/fi/vm/sade/viestintapalvelu/letter/LetterResourceTest.java
@@ -85,7 +85,7 @@ public class LetterResourceTest {
     @Test
     public void startsProcessingLetters() {
         resource.asyncLetter(DocumentProviderTestData.getAsyncLetterBatch()).getEntity();
-        verify(processor).processLetterBatch(any(Integer.class));
+        verify(processor).processLetterBatch(any(Long.class));
     }
     
     @Test
